irusb: refactor headers
- remove 'irusb.h' completely, move all state-related declarations to
  'irusb_state.h'
- rename 'irusb.cpp' to 'irusb_app.cpp'
- remove excess includes from the header; add missing includes to the
  source files to make them self-sufficient include-wise
- move all macro definitions to the header
